Add 'regex' strategy to quick open
