+++ browser/src/Services/Menu/Menu.ts
@@ -65,7 +65,7 @@ export class Menu {
     private _onItemSelected = new Event<any>()
     private _onFilterTextChanged = new Event<string>()
     private _onHide = new Event<void>()
-    private _filterFunction = MenuFilter.regexFilter
+    private _filterFunction = MenuFilter.fuseFilter
 
     public get onHide(): IEvent<void> {
         return this._onHide
