+++ browser/src/Services/Menu/MenuFilter.ts
@@ -13,9 +13,7 @@ import { configuration } from "./../../Services/Configuration"
 
 import { IMenuOptionWithHighlights } from "./Menu"
 
-import { createLetterCountDictionary, LetterCountDictionary } from "./../../UI/components/HighlightText"
-
-const shouldFilterbeCaseSensitive = (searchString: string): boolean => {
+export const shouldFilterbeCaseSensitive = (searchString: string): boolean => {
 
     // TODO: Technically, this makes the reducer 'impure',
     // which is not ideal - need to refactor eventually.
@@ -39,62 +37,6 @@ const shouldFilterbeCaseSensitive = (searchString: string): boolean => {
             return true
         }
     }
-}
-
-export const regexFilter = (options: Oni.Menu.MenuOption[], searchString: string): IMenuOptionWithHighlights[] => {
-    if (!searchString) {
-        const opt = options.map((o) => {
-            return {
-                ...o,
-                detailHighlights: [],
-                labelHighlights: [],
-            }
-        })
-
-        return sortBy(opt, (o) => o.pinned ? 0 : 1)
-    }
-
-    const filterRegExp = new RegExp(".*" + searchString.split("").join(".*") + ".*")
-
-    const filteredOptions = options.filter((f) => {
-        const textToFilterOn = f.detail + f.label
-        return textToFilterOn.match(filterRegExp)
-    })
-
-    const ret = filteredOptions.map((fo) => {
-        const letterCountDictionary = createLetterCountDictionary(searchString)
-
-        const detailHighlights = getHighlightsFromString(fo.detail, letterCountDictionary)
-        const labelHighlights = getHighlightsFromString(fo.label, letterCountDictionary)
-
-        return {
-            ...fo,
-            detailHighlights,
-            labelHighlights,
-        }
-    })
-
-    return ret
-}
-
-export const getHighlightsFromString = (text: string, letterCountDictionary: LetterCountDictionary): number[] => {
-
-    if (!text) {
-        return []
-    }
-
-    const ret: number[] = []
-
-    for (let i = 0; i < text.length; i++) {
-        const letter = text[i]
-        const idx = i
-        if (letterCountDictionary[letter] && letterCountDictionary[letter] > 0) {
-            ret.push(idx)
-            letterCountDictionary[letter]--
-        }
-    }
-
-    return ret
 }
 
 export const fuseFilter = (options: Oni.Menu.MenuOption[], searchString: string): IMenuOptionWithHighlights[] => {
