+++ browser/src/Services/Menu/index.ts
@@ -1,2 +1,3 @@
 export * from "./Menu"
 export * from "./MenuComponent"
+export * from "./MenuFilter"
